Specialize DBeaver folders to be synced
The original list of DBeaver's folders to be synced was too broad and
could sync lots of unecessary files. This would not only make the backup
bigger, but it would also be more prone to create conficts between the
instances of DBeaver sharing the same files.

This commit specify exacly which files and subfolders of the DBeaver's
main configuration folder should be synced.